+++ src/com/csc/training/parse/ParseForm10kReducer.java
@@ -3,22 +3,21 @@ package com.csc.training.parse;
 import java.io.IOException;
 import java.util.Iterator;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
-import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.MapReduceBase;
 import org.apache.hadoop.mapred.OutputCollector;
 import org.apache.hadoop.mapred.Reducer;
 import org.apache.hadoop.mapred.Reporter;
 
-public class ParseForm10kReducer extends MapReduceBase implements Reducer<ParseForm10kMapKey, IntWritable, Text, IntWritable> {
+public class ParseForm10kReducer extends MapReduceBase implements Reducer<ParseForm10kMapKey, FloatWritable, Text, FloatWritable> {
 	
 	@Override
-	public void reduce(ParseForm10kMapKey key, Iterator<IntWritable> values,
-			OutputCollector<Text, IntWritable> output, Reporter reporter) throws IOException {
+	public void reduce(ParseForm10kMapKey key, Iterator<FloatWritable> values,
+			OutputCollector<Text, FloatWritable> output, Reporter reporter) throws IOException {
 			
 		Text reduceKey = new Text();
-		IntWritable reduceValue = new IntWritable();
+		FloatWritable reduceValue = new FloatWritable();
 		int sum = 0;
 		
 		while(values.hasNext()){
